+++ Native/Android/LunarConsole/lunarConsole/src/main/java/spacemadness/com/lunarconsole/console/actions/Action.java
@@ -21,13 +21,6 @@
 
 package spacemadness.com.lunarconsole.console.actions;
 
-import android.content.Context;
-import android.view.View;
-import android.widget.TextView;
-
-import spacemadness.com.lunarconsole.R;
-import spacemadness.com.lunarconsole.console.ConsoleActionAdapter;
-
 public class Action extends BaseIdentityEntry
 {
     public Action(int actionId, String name)
@@ -40,30 +33,4 @@ public class Action extends BaseIdentityEntry
     {
         return EntryType.Action;
     }
-
-    //region ViewHolder
-
-    public static class ViewHolder extends ConsoleActionAdapter.ViewHolder<Action>
-    {
-        private final View layout;
-        private final TextView nameView;
-
-        public ViewHolder(View itemView)
-        {
-            super(itemView);
-
-            layout = itemView.findViewById(R.id.lunar_console_action_entry_layout);
-            nameView = (TextView) itemView.findViewById(R.id.lunar_console_action_entry_name);
-        }
-
-        @Override
-        public void onBindViewHolder(Action action, int position)
-        {
-            Context context = getContext();
-            layout.setBackgroundColor(action.getBackgroundColor(context, position));
-            nameView.setText(action.name());
-        }
-    }
-
-    //endregion
 }
